+++ source/tests/lambdas/cloudformation_resources/test_unique_name.py
@@ -12,6 +12,7 @@
 # #####################################################################################################################
 
 import pytest
+
 from lambdas.cloudformation_resources.unique_name import generate_name, helper