+++ source/tests/lambdas/sns/test_sns.py
@@ -12,12 +12,19 @@
 # #####################################################################################################################
 
 import json
+import logging
 
 import boto3
 import pytest
 from moto import mock_sns
 
-from lambdas.sns.handler import sns, sns_conditional
+from lambdas.sns.handler import (
+    sns,
+    sns_conditional,
+    prepare_forecast_ready_message,
+)
+from shared.Dataset.dataset_file import DatasetFile
+from shared.Dataset.dataset_type import DatasetType
 
 fail_state_error_message = json.loads(
     """
@@ -49,11 +56,35 @@ fail_service_error_message = json.loads(
 """
 )
 
+in_progress_message = json.loads(
+    """
+{
+  "bucket": "some_bucket",
+  "dataset_file": "train/some_forecast_name.csv",
+  "DatasetGroupArn": "arn:aws:forecast:us-east-1:abcdefghijkl:dataset-group/some_forecast_name",
+  "DatasetArn": "arn:aws:forecast:us-east-1:abcdefghijkl:dataset/some_forecast_name",
+  "serviceError": {
+    "Error": "DatasetsImporting",
+    "Cause": "{\\"errorMessage\\": \\"in progress message goes here\\"}"
+  }
+}
+"""
+)
+
 succeed = json.loads(
     """
 {
   "bucket": "some_bucket",
   "dataset_file": "train/some_forecast_name.csv"
+}
+"""
+)
+
+succeed_related = json.loads(
+    """
+{
+  "bucket": "some_bucket",
+  "dataset_file": "train/some_forecast_name.related.csv"
 }
 """
 )
@@ -70,8 +101,18 @@ def fail_service_error():
 
 
 @pytest.fixture
+def in_progress():
+    return in_progress_message
+
+
+@pytest.fixture
 def success_event():
     return succeed
+
+
+@pytest.fixture
+def success_event_related():
+    return succeed_related
 
 
 @pytest.fixture(scope="function")
@@ -80,6 +121,13 @@ def mocked_sns():
         cli = boto3.client("sns", region_name="us-east-1")
         cli.create_topic(Name="some-forecast-notification-topic")
         yield cli
+
+
+@pytest.fixture(params=["dataset", "dataset.related", "dataset.metadata"])
+def dataset_file(request) -> DatasetFile:
+    csv_name = f"train/{request.param}.csv"
+    dsf = DatasetFile(csv_name, "bucketname")
+    return dsf
 
 
 def test_sns_notification(fail_state_error, mocked_sns):
@@ -101,3 +149,19 @@ def test_sns_conditional_error(fail_service_error, mocker):
     patched_client = mocker.patch("lambdas.sns.handler.get_sns_client")
     sns_conditional(fail_service_error_message, None)
     assert patched_client().publish.called
+
+
+def test_sns_notification_in_progress(in_progress, mocker):
+    patched_client = mocker.patch("lambdas.sns.handler.get_sns_client")
+    sns(in_progress, None)
+    patched_client().publish.assert_called_once()
+    assert "Update for forecast" in patched_client().publish.call_args_list[
+        0
+    ].kwargs.get("Message")
+
+
+def test_prepare_forecast_ready_message(dataset_file):
+    assert (
+        prepare_forecast_ready_message(dataset_file)
+        == f"Forecast for {dataset_file.prefix} is ready!"
+    )