+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.1] - 2020-08-07
+### Fixed
+- fixed a known issue causing multiple dataset upload to fail
+- fixed an issue where users would be notified multiple times that a forecast had completed successfully   
+### Changed
+- anonymous solution metrics publish on all stack changes
+- improved update emails as datasets are uploaded
+
+
 ## [1.0.0] - 2020-07-02
 ### Added
 - initial release
+